move application into its own file
